+++ Nickel.Bugfixes/Fixes/ExhaustEntireHandActionFixes.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Reflection.Emit;
 
-namespace Nickel.Essentials;
+namespace Nickel.Bugfixes;
 
 internal static class ExhaustEntireHandActionFixes
 {