+++ Nickel.Bugfixes/Fixes/IsaacUnlockFixes.cs
@@ -2,7 +2,7 @@ using HarmonyLib;
 using System;
 using System.Reflection;
 
-namespace Nickel.Essentials;
+namespace Nickel.Bugfixes;
 
 internal static class IsaacUnlockFixes
 {