+++ docs/release-notes.md
@@ -7,6 +7,7 @@
 ### For everyone:
 * Browsing through cards in hand for action selection will now display the cards in their order in hand.
 * Fixed various bugs that could happen if card data depends on other cards.
+* Split off a new internal mod `Nickel.Bugfixes` out of `Nickel.Essentials`.
 
 ### For developers:
 * Added the `Pool` and `MultiPool` utilities.